[util] Split up Modes.py and rename
The capitalised filenames are named after classes (and I have just
renamed "Modes" to "Mode"). Also, this file defined lots of different
classes.

This commit pulls out the more interesting subclasses, keeping with
the CapitalisedName style and then renames the remaining file to
modes.py so that we're no longer claiming anything to do with a class
called Modes.
